Level updates plus Wall Class
The level reads in the stuff correctly so far, but it does nothing
with enemies yet until we get ai patterns going. Also, the level object
only draws 2 walls out of the 5 in its walls vector and i have no idea
why. Im hoping that fresh eyes will be able to figure it out.